refactor code directory
