+++ app/src/main/java/com/example/freightcrayt/activities/add_category.java
@@ -1,8 +1,10 @@
-package com.example.freightcrayt;
+package com.example.freightcrayt.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.freightcrayt.R;
 
 public class add_category extends AppCompatActivity {
